requirements.txt
google-generativeai
dotenv
tabulate
rich
json